+++ include/cocotte/models/node.h
@@ -24,7 +24,7 @@ class Node : public Model
 private:
 
     std::vector<Approximators::Form> forms;
-    int nbPoints;
+    unsigned int nbPoints;
     double biggestInnerDistance = -1.;
     std::shared_ptr<Model> model0;
     std::shared_ptr<Model> model1;
@@ -37,8 +37,8 @@ public:
     Node(std::shared_ptr<Model> model0, std::shared_ptr<Model> model1, bool temporary = false);
     virtual bool isLeaf() const override;
     virtual bool isTemporary() const override;
-    virtual size_t getNbPoints() const override;
-    double getBiggestInnerDistance(int outputID);   // Biggest distance between two merged submodels
+    virtual unsigned int getNbPoints() const override;
+    double getBiggestInnerDistance(unsigned int outputID);   // Biggest distance between two merged submodels
     virtual std::vector<Approximators::Form> const& getForms() override;
     void setForms(std::vector<Approximators::Form> const& forms);
